requirements.txt
PyCryptodome

pip
setuptools